dagster/requirements.txt
# ... existing code ...
psycopg2-binary>=2.9.3
